Hydra.Contract.Commit

Description

Datum and redeemer types, as well as helper functions for the commit validator implemented in aiken.

Documentation

data Commit

A data type representing comitted outputs on-chain. Besides recording the original TxOutRef, it also stores a binary representation compatible between on- and off-chain code to be hashed in the validators.

Constructors

Commit

Fields

input :: TxOutRef

preSerializedOutput :: BuiltinByteString

serializeCommit :: (TxIn, TxOut CtxUTxO) -> Maybe Commit

Record an off-chain TxOut as a Commit on-chain. NOTE: Depends on the Serialise instance for Plutus' Data.

deserializeCommit :: Network -> Commit -> Maybe (TxIn, TxOut CtxUTxO)

Decode an on-chain SerializedTxOut back into an off-chain TxOut. NOTE: Depends on the Serialise instance for Plutus' Data.

type DatumType = (Party, [Commit], CurrencySymbol)

data CommitRedeemer

Constructors

ViaCollectCom

ViaAbort

type RedeemerType = CommitRedeemer

datum :: DatumType -> Datum

redeemer :: RedeemerType -> Redeemer