Hydra.Contract.Deposit

Description

The validator used to deposit and recover locked funds

Documentation

data DepositRedeemer

Constructors

Claim
Claims already deposited funds.

Recover Integer
Recovers m number of deposited outputs.

newtype DepositDatum

Deposit datum containing HeadId, deadline and a list of deposits.

Constructors

DepositDatum (CurrencySymbol, POSIXTime, [Commit])

validator :: DepositDatum -> DepositRedeemer -> ScriptContext -> Bool

v_deposit validator checks

Claim redeemer -> more checks will be added

Recover redeemer

The deadline has been reached.

The hash of recovered outputs are matching the deposited outputs.

compiledValidator :: CompiledCode ValidatorType

validatorScript :: SerialisedScript

validatorHash :: ScriptHash

datum :: DepositDatum -> Datum

redeemer :: DepositRedeemer -> Redeemer