Hydra.Contract.Hash

Description

An experimental validator which simply hashes a bytestring stored in the datum using one of three supported algorithms.

Documentation

data HashAlgorithm

Constructors

Base

SHA2

SHA3

Blake2b

type DatumType = BuiltinByteString

type RedeemerType = HashAlgorithm

validator :: DatumType -> RedeemerType -> ScriptContext -> Bool

compiledValidator :: CompiledCode ValidatorType

validatorScript :: SerialisedScript

validatorHash :: ScriptHash

datum :: DatumType -> Datum

redeemer :: RedeemerType -> Redeemer