Hydra.Contract.HeadState

Documentation

type SnapshotNumber = Integer

type SnapshotVersion = Integer

type Hash = BuiltinByteString

type Signature = BuiltinByteString

data OpenDatum

Sub-type for the open state-machine state.

Constructors

OpenDatum

Fields

headId :: CurrencySymbol

Spec: cid

parties :: [Party]

Spec: kH

contestationPeriod :: ContestationPeriod

Spec: T

version :: SnapshotVersion

Spec: v

utxoHash :: Hash

Spec: η

data ClosedDatum

Sub-type for the closed state-machine state.

Constructors

ClosedDatum

Fields

headId :: CurrencySymbol

Spec: cid

parties :: [Party]

Spec: kH

contestationPeriod :: ContestationPeriod

Spec: T

version :: SnapshotVersion

Spec: v

snapshotNumber :: SnapshotNumber

Spec: s

utxoHash :: Hash

Spec: η. Digest of snapshotted UTxO

alphaUTxOHash :: Hash

omegaUTxOHash :: Hash

Spec: ηΔ. Digest of UTxO still to be distributed

contesters :: [PubKeyHash]

Spec: C

contestationDeadline :: POSIXTime

Spec: tfinal

data State

Constructors

Initial

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

headId :: CurrencySymbol

seed :: TxOutRef

Open OpenDatum

Closed ClosedDatum

Final

data CloseRedeemer

Sub-type for close transition with auxiliary data as needed.

Constructors

CloseInitial
Initial snapshot is used to close.

CloseAny
Any snapshot which doesn't contain anything to inc/decrement but snapshot number is higher than zero.

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

CloseUnusedDec
Closing snapshot refers to the current state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

CloseUsedDec
Closing snapshot refers to the previous state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

alreadyDecommittedUTxOHash :: Hash

UTxO which was already decommitted ηω

CloseUnusedInc
Closing snapshot refers to the current state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

alreadyCommittedUTxOHash :: Hash

UTxO which was signed but not committed ηα

CloseUsedInc
Closing snapshot refers to the previous state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

alreadyCommittedUTxOHash :: Hash

UTxO which was signed but not committed ηα

data ContestRedeemer

Sub-type for contest transition with auxiliary data as needed.

Constructors

ContestCurrent
Contesting snapshot refers to the current state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

ContestUsedDec
Contesting snapshot refers to the previous state version

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

alreadyDecommittedUTxOHash :: Hash

UTxO which was already decommitted ηω

ContestUnusedDec
Redeemer to use when the decommit was not yet observed but we closed the Head.

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

ContestUnusedInc
Redeemer to use when the commit was not yet observed but we closed the Head.

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

alreadyCommittedUTxOHash :: Hash

UTxO which was already committed ηα

ContestUsedInc

Fields

signature :: [Signature]

Multi-signature of a snapshot ξ

data IncrementRedeemer

Sub-type for increment transition

Constructors

IncrementRedeemer

Fields

signature :: [Signature]

snapshotNumber :: SnapshotNumber

increment :: TxOutRef

data DecrementRedeemer

Sub-type for decrement transition with auxiliary data as needed.

Constructors

DecrementRedeemer

Fields

signature :: [Signature]

Spec: ξ

snapshotNumber :: SnapshotNumber

Spec: s

numberOfDecommitOutputs :: Integer

Spec: m

data Input

Constructors

CollectCom

Increment IncrementRedeemer

Decrement DecrementRedeemer

Close CloseRedeemer

Contest ContestRedeemer

Abort

Fanout

Fields

numberOfFanoutOutputs :: Integer

numberOfCommitOutputs :: Integer

numberOfDecommitOutputs :: Integer