Hydra.Contract.HeadTokens

Description

Minting policy for a single head tokens.

Documentation

validate

Arguments

:: ValidatorHash
Head validator

-> ValidatorHash

-> TxOutRef

-> MintAction

-> ScriptContext

-> Bool

validateTokensMinting :: ValidatorHash -> ValidatorHash -> TxOutRef -> ScriptContext -> Bool

When minting head tokens we want to make sure that:

The number of minted PTs == number of participants (+1 for the ST) evident from the datum.

There is single state token that is paid into v_head, which ensures continuity.

PTs are distributed to v_initial.

Ensure out-ref and the headId are in the datum of the first output of the transaction which mints tokens.

validateTokensBurning :: ScriptContext -> Bool

Token burning check should: * Not restrict burning on the mu_head at all.

It is ensured by the v_head validator, when tokens of a specific headId may be burned.

validateTokensBurning just makes sure all tokens have negative quantity.

unappliedMintingPolicy :: CompiledCode (TxOutRef -> MintingPolicyType)

Raw minting policy code where the TxOutRef is still a parameter.

mintingPolicyScript :: TxOutRef -> Script

Get the applied head minting policy script given a seed TxOutRef.

Create PolicyId

headPolicyId :: TxIn -> PolicyId

Get the head policy id (a.k.a headId) given a seed TxIn.

mkHeadTokenScript :: TxIn -> PlutusScript

Get the applied head minting policy script given a seed TxIn.