Hydra.Contract.Initial

Description

The Initial validator which allows participants to commit or abort. To focus on the off-chain datum types this is currently 'const True'.

Documentation

data Initial

data InitialRedeemer

Constructors

Abort

Commit

Fields

committedRef :: Maybe TxOutRef

Points to the committed Utxo.

validator

Arguments

:: ValidatorHash
Commit validator

-> ()

-> InitialRedeemer

-> ScriptContext

-> Bool

The initial validator has two responsibilities:

ensures the committed value is recorded correctly in the output datum

ensures that the transaction was signed by the key corresponding to the PubKeyHash encoded in the participation token name

NOTE: It does not need to ensure that the participation token is of some specific Head currency.

checkAuthor :: ScriptContext -> Bool

Verifies that the commit is only done by the author

checkCommit

Arguments

:: ValidatorHash
Commit validator

-> Maybe TxOutRef

-> ScriptContext

-> Bool

debugInteger :: Integer -> BuiltinString

Show an Integer as decimal number. This is very inefficient and only should be used for debugging.

typedValidator :: TypedValidator Initial

compiledValidator :: CompiledCode (ValidatorType Initial)

validatorScript :: Script

Get the actual plutus script. Mainly used to serialize and use in transactions.

validatorHash :: ValidatorHash

datum :: DatumType Initial -> Datum

redeemer :: RedeemerType Initial -> Redeemer

address :: Address