Hydra.Contract.OffChain

Documentation

data HeadParameters

Constructors

HeadParameters

Fields

participants :: [PubKeyHash]

policy :: MonetaryPolicy

policyId :: MonetaryPolicyHash

mkHeadParameters :: [PubKeyHash] -> MonetaryPolicy -> HeadParameters

init :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

commit :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

collectCom :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

abort :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

setupForTesting :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

This endpoint allows for setting up a wallet for testing, that is, a wallet that has several UTxO, so that one can be locked and the other used to pay for fees.

type Schema = ((((BlockchainActions .\/ Endpoint "setupForTesting" (PubKeyHash, [Value])) .\/ Endpoint "init" ()) .\/ Endpoint "commit" (PubKeyHash, (TxOutRef, TxOutTx))) .\/ Endpoint "collectCom" (PubKeyHash, [TxOut])) .\/ Endpoint "abort" (PubKeyHash, [TxOut])

contract :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

data SomeScriptInstance

Constructors

forall a. SomeScriptInstance (ScriptInstance a)

hydraScriptInstance :: HeadParameters -> ScriptInstance Hydra

initialScriptInstance :: HeadParameters -> ScriptInstance Initial

commitScriptInstance :: HeadParameters -> ScriptInstance Commit

toOnChainHeadParameters :: HeadParameters -> HeadParameters

flattenUtxo :: UtxoMap -> [(TxOutRef, TxOut)]

mustRefund :: forall i o. TxOut -> TxConstraints i o

utxoAtWithDatum :: forall w s e. (AsContractError e, HasUtxoAt s) => Address -> Datum -> Contract w s e UtxoMap

zipOnParty :: MonetaryPolicyHash -> [PubKeyHash] -> UtxoMap -> [(PubKeyHash, TxOutRef)]