Hydra.Contract.OnChain

Documentation

data HeadParameters

Constructors

HeadParameters

Fields

participants :: [PubKeyHash]

policyId :: MintingPolicyHash

data State

Constructors

Initial

Open [TxOut]

Final

data Transition

Constructors

CollectCom

Abort

data Hydra

hydraValidator :: HeadParameters -> State -> Transition -> ScriptContext -> Bool

hydraTypedValidator :: HeadParameters -> TypedValidator Hydra

hydraValidatorHash :: HeadParameters -> ValidatorHash

data Initial

initialValidator :: HeadParameters -> ValidatorHash -> ValidatorHash -> PubKeyHash -> TxOutRef -> ScriptContext -> Bool

The Validator initial ensures that the input is consumed by a commit or an abort transaction.

initialTypedValidator :: HeadParameters -> TypedValidator Initial

data Commit

commitValidator :: ValidatorHash -> TxOut -> () -> ScriptContext -> Bool

To lock outputs for a Hydra head, the ith head member will attach a commit transaction to the i-th output of the initial transaction. Validator commit ensures that the commit transaction correctly records the partial UTxO set Ui committed by the party.

The data field of the output of the commit transaction is

Ui = makeUTxO(o_1 , . . . , o_m),

where the o_j are the outputs referenced by the commit transaction’s inputs and makeUTxO stores pairs (out-ref_j , o_j) of outputs o_j with the corresponding output reference out-ref_j .

commitTypedValidator :: HeadParameters -> TypedValidator Commit

commitValidatorHash :: HeadParameters -> ValidatorHash

type FakeTxOutRef = Integer

validateMintingPolicy :: FakeTxOutRef -> () -> ScriptContext -> Bool

hydraMintingPolicy :: FakeTxOutRef -> MintingPolicy

asDatum :: ToData a => a -> Datum

This function can be rather unsafe and should *always* be used with a type annotation to avoid silly mistakes and hours of debugging.

So prefer:

>>> asDatum @(DatumType MyContract) val

over

>>> asDatum val

A Datum is an opaque type, and type information is lost when turning into a a Data. Hence, various functions of the Plutus Tx framework that rely on Datum or Redeemer are basically stringly-typed function with no guarantee whatsoever that the 'IsData a' being passed will correspond to what's expected by the underlying script. So, save you some hours of debugging and always explicitly specify the source type when using this function, preferably using the data-family from the ValidatorTypes instance.

asRedeemer :: ToData a => a -> Redeemer

Always use with explicit type-annotation, See warnings on asDatum.

mustBeSignedByOneOf :: [PubKeyHash] -> ScriptContext -> Bool

mustReimburse :: TxOut -> ScriptContext -> Bool

mustRunContract :: forall redeemer. ToData redeemer => ValidatorHash -> redeemer -> ScriptContext -> Bool

mustForwardParty :: ScriptContext -> MintingPolicyHash -> PubKeyHash -> Bool

mustBurnParty :: ScriptContext -> MintingPolicyHash -> PubKeyHash -> Bool

mkParty :: MintingPolicyHash -> PubKeyHash -> Value

mkPartyName :: PubKeyHash -> TokenName

hasParty :: MintingPolicyHash -> TxInInfo -> Bool

filterInputs :: (TxInInfo -> Bool) -> ScriptContext -> [(TxOutRef, TxOut)]

findUtxo :: TxOutRef -> ScriptContext -> Maybe (TxOutRef, TxOut)

findContractInput :: ValidatorHash -> ScriptContext -> Maybe TxOutRef

symbols :: Value -> [CurrencySymbol]