Hydra.Contract.PAB

Documentation

pabPort :: Int

Hard-coded port used between hydra-node and the PAB server.

data PabContract

Enumeration of contracts available in the PAB.

Constructors

GetUtxos

Init

Watch

getUtxo :: Contract (Last (Map TxOutRef TxOut)) Empty ContractError ()

data InitParams

Parameters for starting a head. NOTE: kinda makes sense to have them separate because it couuld be the case that the parties (hdyra nodes taking part in the head consensus) and th participants (people commiting UTxOs) and posting transaction in the head are different

Constructors

InitParams

Fields

contestationPeriod :: ContestationPeriod

cardanoPubKeys :: [PubKeyHash]

hydraParties :: [Party]

threadTokenName :: TokenName

participationTokenName :: PubKeyHash -> TokenName

mkThreadToken :: CurrencySymbol -> AssetClass

init :: Promise () (Endpoint "init" InitParams) HydraPlutusError ()

data ObservedTx

Transactions as they are observed by the PAB. We use a distinct type from the Chain as we want to keep hydra-node not tied to hydra-plutus. Tests are ensuring that this type as the same wire format as OnChainTx.

Constructors

OnInitTx

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

OnAbortTx

watch :: Contract (Last ObservedTx) (Endpoint "abort" ()) HydraPlutusError ()

watchForSM :: AssetClass -> Contract (Last ObservedTx) s HydraPlutusError ()

watchInit :: Contract w s HydraPlutusError (ContestationPeriod, [Party], AssetClass)

Watch Initial script address to extract head's initial parameters and thread token. Relies on the fact that a participation token (payed to the address) uses the same CurrencySymbol as the thread token.

abort :: AssetClass -> Promise w (Endpoint "abort" ()) HydraPlutusError ()

data HydraPlutusError

Constructors

SMError SMContractError
State machine operation failed

PlutusError ContractError
Endpoint, coin selection, etc. failed

ThreadTokenError CurrencyError
Thread token could not be created

HydraError String
Arbitrary error

class AsHydraPlutusError r where

Minimal complete definition

_HydraPlutusError

Methods

_HydraPlutusError :: Prism' r HydraPlutusError

_SMError :: Prism' r SMContractError

_PlutusError :: Prism' r ContractError

_ThreadTokenError :: Prism' r CurrencyError

_HydraError :: Prism' r String