Hydra.ContractSM

Description

Simplified SM-based contract for the purpose of developing the interface between Node and Chain

Documentation

data State

Constructors

Setup

Initial ContestationPeriod [Party]

Open

Final

data Input

Constructors

Init ContestationPeriod [(PubKeyHash, Value)] [Party]

CollectCom

Abort

data HydraPlutusError

Constructors

SMError SMContractError
State machine operation failed

PlutusError ContractError
Endpoint, coin selection, etc. failed

ThreadTokenError CurrencyError
Thread token could not be created

HydraError String
Arbitrary error

class AsHydraPlutusError r where

Minimal complete definition

_HydraPlutusError

Methods

_HydraPlutusError :: Prism' r HydraPlutusError

_SMError :: Prism' r SMContractError

_PlutusError :: Prism' r ContractError

_ThreadTokenError :: Prism' r CurrencyError

_HydraError :: Prism' r String

hydraStateMachine :: AssetClass -> StateMachine State Input

hydraTransition :: State State -> Input -> Maybe (TxConstraints Void Void, State State)

typedValidator :: AssetClass -> TypedValidator (StateMachine State Input)

The script instance of the auction state machine. It contains the state machine compiled to a Plutus core validator script. The AssetClass serves two roles here:

Parameterizing the script, such that we get a unique address and allow for multiple instances of it

Identify the 'state thread token', which should be passed in transactions transitioning the state machine and provide "contract continuity"

machineClient

Arguments

:: AssetClass
Thread token of the instance

-> StateMachineClient State Input

The machine client of the hydra state machine. It contains both, the script instance with the on-chain code, and the Haskell definition of the state machine for off-chain use.

participationTokenName :: PubKeyHash -> TokenName

threadTokenName :: TokenName

mkThreadToken :: CurrencySymbol -> AssetClass

data InitParams

Parameters for starting a head. NOTE: kinda makes sense to have them separate because it couuld be the case that the parties (hdyra nodes taking part in the head consensus) and th participants (people commiting UTxOs) and posting transaction in the head are different

Constructors

InitParams

Fields

contestationPeriod :: ContestationPeriod

cardanoPubKeys :: [PubKeyHash]

hydraParties :: [Party]

setup :: Contract () (Endpoint "init" InitParams) HydraPlutusError ()

data InitialParams

Parameters as they are available in the Initial state.

Constructors

InitialParams

Fields

contestationPeriod :: ContestationPeriod

parties :: [Party]

watchInit :: Contract (Last InitialParams) Empty ContractError ()

Watch initialAddress (with hard-coded parameters) and report all datums seen on each run.

watchStateMachine :: AssetClass -> Contract () Empty HydraPlutusError State

Wait for Init transaction to appear on chain and return the observed state of the state machine