Hydra.Depreciated.OffChain

Documentation

data HeadParameters

Constructors

HeadParameters

Fields

participants :: [PubKeyHash]

policy :: MintingPolicy

policyId :: MintingPolicyHash

mkHeadParameters :: [PubKeyHash] -> MintingPolicy -> HeadParameters

init :: AsContractError e => HeadParameters -> Promise [State] Schema e ()

commit :: AsContractError e => HeadParameters -> Promise [State] Schema e ()

collectCom :: AsContractError e => HeadParameters -> Promise [State] Schema e ()

abort :: AsContractError e => HeadParameters -> Promise [State] Schema e ()

setupForTesting :: AsContractError e => HeadParameters -> Promise [State] Schema e ()

This endpoint allows for setting up a wallet for testing, that is, a wallet that has several UTxO, so that one can be locked and the other used to pay for fees.

type Schema = (((Endpoint "setupForTesting" (PubKeyHash, [Value]) .\/ Endpoint "init" ()) .\/ Endpoint "commit" (PubKeyHash, (TxOutRef, TxOut))) .\/ Endpoint "collectCom" (PubKeyHash, [TxOut])) .\/ Endpoint "abort" (PubKeyHash, [TxOut])

contract :: AsContractError e => HeadParameters -> Contract [State] Schema e ()

data SomeTypedValidator

Constructors

forall a. SomeTypedValidator (TypedValidator a)

hydraTypedValidator :: HeadParameters -> TypedValidator Hydra

initialTypedValidator :: HeadParameters -> TypedValidator Initial

commitTypedValidator :: HeadParameters -> TypedValidator Commit

toOnChainHeadParameters :: HeadParameters -> HeadParameters

mustRefund :: forall i o. TxOut -> TxConstraints i o

utxoAt :: forall w s e. AsContractError e => Address -> Contract w s e (Map TxOutRef TxOut)

Thin wrapper around utxosAt to just get a 'Map TxOutRef TxOut'. Created to avoid corruptions into ChainIndexTxOut.

utxoAtWithDatum :: forall w s e. AsContractError e => Address -> Datum -> Contract w s e (Map TxOutRef TxOut)

zipOnParty :: MintingPolicyHash -> [PubKeyHash] -> Map TxOutRef TxOut -> [(PubKeyHash, TxOutRef)]