Hydra.OnChain.Util

Documentation

asDatum :: ToData a => a -> Datum

This function can be rather unsafe and should *always* be used with a type annotation to avoid silly mistakes and hours of debugging.

So prefer:

>>> asDatum @(DatumType MyContract) val

over

>>> asDatum val

A Datum is an opaque type, and type information is lost when turning into a a Data. Hence, various functions of the Plutus Tx framework that rely on Datum or Redeemer are basically stringly-typed function with no guarantee whatsoever that the 'IsData a' being passed will correspond to what's expected by the underlying script. So, save you some hours of debugging and always explicitly specify the source type when using this function, preferably using the data-family from the ValidatorTypes instance.

asRedeemer :: ToData a => a -> Redeemer

Always use with explicit type-annotation, See warnings on asDatum.

mustBeSignedByOneOf :: [PubKeyHash] -> ScriptContext -> Bool

mustReimburse :: TxOut -> ScriptContext -> Bool

mustRunContract :: forall redeemer. ToData redeemer => ValidatorHash -> redeemer -> ScriptContext -> Bool

mustForwardParty :: ScriptContext -> MintingPolicyHash -> PubKeyHash -> Bool

mustBurnParty :: ScriptContext -> MintingPolicyHash -> PubKeyHash -> Bool

mkParty :: MintingPolicyHash -> PubKeyHash -> Value

mkPartyName :: PubKeyHash -> TokenName

hasParty :: MintingPolicyHash -> TxInInfo -> Bool

filterInputs :: (TxInInfo -> Bool) -> ScriptContext -> [(TxOutRef, TxOut)]

findUtxo :: ScriptContext -> TxOutRef -> Maybe (TxOutRef, TxOut)

findContractInput :: ValidatorHash -> ScriptContext -> Maybe TxOutRef

symbols :: Value -> [CurrencySymbol]