Hydra.ScriptContext

Description

A custom ScriptContext and TxInfo which only "decodes" the fields we need.

Tx info

data TxInfo

Constructors

TxInfo

Fields

txInfoInputs :: [TxInInfo]

Transaction inputs; cannot be an empty list

txInfoReferenceInputs :: BuiltinData

Transaction reference inputs

txInfoOutputs :: [TxOut]

Transaction outputs

txInfoFee :: Value

The fee paid by this transaction.

txInfoMint :: Value

The Value minted by this transaction.

txInfoDCert :: BuiltinData

Digests of certificates included in this transaction

txInfoWdrl :: BuiltinData

Withdrawals

txInfoValidRange :: BuiltinData

The valid range for the transaction.

txInfoSignatories :: [PubKeyHash]

Signatures provided with the transaction, attested that they all signed the tx

txInfoRedeemers :: BuiltinData

A table of redeemers attached to the transaction

txInfoData :: Map DatumHash Datum

The lookup table of datums attached to the transaction

txInfoId :: BuiltinData

Hash of the pending transaction body (i.e. transaction excluding witnesses)

Script context

data ScriptContext

The context that the currently-executing script can access.

Constructors

ScriptContext

Fields

scriptContextTxInfo :: TxInfo

information about the transaction the currently-executing script is included in

scriptContextPurpose :: ScriptPurpose

the purpose of the currently-executing script

Utilities

scriptOutputsAt :: ScriptHash -> TxInfo -> [(OutputDatum, Value)]

Get the list of TxOut outputs of the pending transaction at a given script address.

valueLockedBy :: TxInfo -> ScriptHash -> Value

Get the total value locked by the given validator in this transaction.

findOwnInput :: ScriptContext -> Maybe TxInInfo

Find the input currently being validated.

findDatum :: DatumHash -> TxInfo -> Maybe Datum

Find the data corresponding to a data hash, if there is one

findTxInByTxOutRef :: TxOutRef -> TxInfo -> Maybe TxInInfo

Given a UTXO reference and a transaction (TxInfo), resolve it to one of the transaction's inputs (TxInInfo).

ownCurrencySymbol :: ScriptContext -> CurrencySymbol

The CurrencySymbol of the current validator script.