Plutus.Contract.StateMachine.OnChain

Description

On-chain code fragments for creating a state machine. First define a StateMachine s i with input type i and state type s. Then use mkValidator in on-chain code to check the required hashes and validate the transition, and mkRedeemer to make redeemer scripts.

Documentation

data StateMachine s i

Specification of a state machine, consisting of a transition function that determines the next state from the current state and an input, and a checking function that checks the validity of the transition in the context of the current transaction.

Constructors

StateMachine

Fields

smTransition :: State s -> i -> Maybe (TxConstraints Void Void, State s)

The transition function of the state machine. Nothing indicates an invalid transition from the current state.

smFinal :: s -> Bool

Check whether a state is the final state

smCheck :: s -> i -> ScriptContext -> Bool

The condition checking function. Can be used to perform checks on the pending transaction that aren't covered by the constraints. smCheck is always run in addition to checking the constraints, so the default implementation always returns true.

smThreadToken :: Maybe ThreadToken

The ThreadToken that identifies the contract instance. Make one with getThreadToken and pass it on to mkStateMachine. Initialising the machine will then mint a thread token value.

data StateMachineInstance s i

Constructors

StateMachineInstance

Fields

stateMachine :: StateMachine s i

The state machine specification.

typedValidator :: TypedValidator (StateMachine s i)

The validator code for this state machine.

data State s

Constructors

State

Fields

stateData :: s

stateValue :: Value

mkStateMachine :: Maybe ThreadToken -> (State s -> i -> Maybe (TxConstraints Void Void, State s)) -> (s -> Bool) -> StateMachine s i

A state machine that does not perform any additional checks on the ScriptContext (beyond enforcing the constraints)

machineAddress :: StateMachineInstance s i -> Address

mkValidator :: forall s i. ToData s => StateMachine s i -> ValidatorType (StateMachine s i)

Turn a state machine into a validator script.

threadTokenValueOrZero :: StateMachineInstance s i -> Value

The Value containing exactly the thread token, if one has been specified.