Plutus.Contract.StateMachine.ThreadToken

Description

Thread token data type definition and minting policy. Thread tokens are used to identify the contract instance on the blockchain, and ensuring that the state was produced by running the state machine from its initial state.

Documentation

data ThreadToken

Constructors

ThreadToken

Fields

ttOutRef :: TxOutRef

ttCurrencySymbol :: CurrencySymbol

checkPolicy :: TxOutRef -> (ValidatorHash, MintingPolarity) -> ScriptContext -> Bool

curPolicy :: TxOutRef -> MintingPolicy

threadTokenValue :: CurrencySymbol -> ValidatorHash -> Value

The Value containing exactly the thread token.

checkThreadTokenInner :: CurrencySymbol -> ValidatorHash -> Value -> Integer -> Bool

checkThreadToken :: Maybe ThreadToken -> ValidatorHash -> Value -> Integer -> Bool