Plutus.Extras

Vendored from plutus-ledger

type ValidatorType = BuiltinData -> BuiltinData -> BuiltinData -> ()

Signature of an untyped validator script.

wrapValidator :: (UnsafeFromData datum, UnsafeFromData redeemer, UnsafeFromData context) => (datum -> redeemer -> context -> Bool) -> ValidatorType

Wrap a typed validator to get the basic ValidatorType signature which can be passed to compile. REVIEW: There might be better ways to name this than "wrap"

type MintingPolicyType = BuiltinData -> BuiltinData -> ()

Signature of an untyped minting policy script.

wrapMintingPolicy :: (UnsafeFromData redeemer, UnsafeFromData context) => (redeemer -> context -> Bool) -> MintingPolicyType

Wrap a typed minting policy to get the basic MintingPolicyType signature which can be passed to compile.

Similar utilities as plutus-ledger

scriptValidatorHash :: Script -> ValidatorHash

Compute the ValidatorHash for a given plutus Script.

NOTE: Implemented using hydra-cardano-api (PlutusScript pattern)