Hydra.Contract.Head

Documentation

type DatumType = State

type RedeemerType = Input

headValidator :: State -> Input -> ScriptContext -> Bool

hashPreSerializedCommits :: [Commit] -> BuiltinByteString

Hash a potentially unordered list of commits by sorting them, concatenating their preSerializedOutput bytes and creating a SHA2_256 digest over that.

hashTxOuts :: [TxOut] -> BuiltinByteString

Hash a pre-ordered list of transaction outputs by serializing each individual TxOut, concatenating all bytes together and creating a SHA2_256 digest over that.

verifyPartySignature :: SnapshotNumber -> BuiltinByteString -> Party -> Signature -> Bool

verifySnapshotSignature :: [Party] -> SnapshotNumber -> BuiltinByteString -> [Signature] -> Bool

compiledValidator :: CompiledCode ValidatorType

validatorScript :: Script

validatorHash :: ValidatorHash