Hydra.Contract.Head

Documentation

type DatumType = State

type RedeemerType = Input

headValidator :: State -> Input -> ScriptContext -> Bool

checkAbort :: ScriptContext -> CurrencySymbol -> [Party] -> Bool

On-Chain verification for Abort transition. It verifies that:

All PTs have been burnt: The right number of Head tokens with the correct head id are burnt, one PT for each party and a state token ST.

All committed funds have been redistributed. This is done via v_commit and it only needs to ensure that we have spent all committed outputs, which follows from burning all the PTs.

checkCollectCom

Arguments

:: ScriptContext
Script execution context

-> (ContestationPeriod, [Party], CurrencySymbol)

-> Bool

On-Chain verification for CollectCom transition. It verifies that:

All participants have committed (even empty commits)

All commits are properly collected and locked into η as a hash of serialized tx outputs in the same sequence as commit inputs!

The transaction is performed (i.e. signed) by one of the head participants

State token (ST) is present in the output

commitDatum :: TxOut -> [Commit]

Try to find the commit datum in the input and if it is there return the committed utxo

depositDatum :: TxOut -> [Commit]

Try to find the deposit datum in the input and if it is there return the committed utxo

checkIncrement

Arguments

:: ScriptContext

-> OpenDatum
Open state before the increment

-> IncrementRedeemer

-> Bool

Verify a increment transaction.

checkDecrement

Arguments

:: ScriptContext

-> OpenDatum
Open state before the decrement

-> DecrementRedeemer

-> Bool

Verify a decrement transaction.

checkClose

Arguments

:: ScriptContext

-> OpenDatum
Open state before the close

-> CloseRedeemer
Type of close transition.

-> Bool

Verify a close transaction.

checkContest

Arguments

:: ScriptContext

-> ClosedDatum
Closed state before the close

-> ContestRedeemer
Type of contest transition.

-> Bool

Verify a contest transaction.

checkFanout

Arguments

:: ScriptContext

-> ClosedDatum
Closed state before the fanout

-> Integer
Number of normal outputs to fanout

-> Integer
Number of alpha outputs to fanout

-> Integer
Number of delta outputs to fanout

-> Bool

Verify a fanout transaction.

makeContestationDeadline :: ContestationPeriod -> ScriptContext -> POSIXTime

getHeadInput :: ScriptContext -> TxInInfo

This is safe only because usually Head transaction only consume one input.

getHeadAddress :: ScriptContext -> Address

mustNotChangeParameters :: ([Party], [Party]) -> (ContestationPeriod, ContestationPeriod) -> (CurrencySymbol, CurrencySymbol) -> Bool

mustBeSignedByParticipant :: ScriptContext -> CurrencySymbol -> Bool

findParticipationTokens :: CurrencySymbol -> Value -> [TokenName]

headOutputDatum :: ScriptContext -> Datum

getTxOutDatum :: TxOut -> Datum

hasPT :: CurrencySymbol -> TxOut -> Bool

Check if TxOut contains the PT token.

verifySnapshotSignature :: [Party] -> (CurrencySymbol, SnapshotVersion, SnapshotNumber, Hash, Hash, Hash) -> [Signature] -> Bool

Verify the multi-signature of a snapshot using given constituents State, ClosedDatum, number, ClosedDatum and utxoToDecommitHash. See 'SignableRepresentation Snapshot' for more details.

verifyPartySignature :: (CurrencySymbol, SnapshotVersion, SnapshotNumber, Hash, Hash, Hash) -> Party -> Signature -> Bool

Verify individual party signature of a snapshot. See 'SignableRepresentation Snapshot' for more details.

compiledValidator :: CompiledCode ValidatorType

validatorScript :: PlutusScript

decodeHeadOutputClosedDatum :: ScriptContext -> ClosedDatum

decodeHeadOutputOpenDatum :: ScriptContext -> OpenDatum

emptyHash :: Hash