Hydra.Contract.Head

Documentation

data Head

hydraHeadV1 :: BuiltinByteString

headValidator

Arguments

:: Address
Commit script address. NOTE: Used to identify inputs from commits and likely could be replaced by looking for PTs.

-> Address
Inital script address. NOTE: Used to identify inputs from initials and likely could be replaced by looking for PTs.

-> State

-> Input

-> ScriptContext

-> Bool

data CheckCollectComError

Constructors

NoContinuingOutput

MoreThanOneContinuingOutput

OutputValueNotPreserved

OutputHashNotMatching

data HeadContext

Constructors

HeadContext

Fields

headAddress :: Address

headInputValue :: Value

headCurrencySymbol :: CurrencySymbol

commitAddress :: Address

initialAddress :: Address

mkHeadContext :: ScriptContext -> Address -> Address -> HeadContext

checkAbort :: ScriptContext -> HeadContext -> [Party] -> Bool

On-Chain Validation for Abort transition. It must verify that: * All PTs have been burnt * It has collected inputs for all parties, either from Initial or Commit script.

checkCollectCom

Arguments

:: ScriptContext
Script execution context

-> HeadContext
Static information about the head (i.e. address, value, currency...)

-> (ContestationPeriod, [Party])
Initial state

-> Bool

On-Chain Validation for the CollectCom transition.

The CollectCom transition must verify that:

All participants have committed (even empty commits)

All commits are properly collected and locked into the contract

The transaction is performed (i.e. signed) by one of the head participants

It must also Initialize the on-chain state η* with a snapshot number and a Merkle-Tree root hash of committed outputs.

(*) In principle, η contains not a hash but a full UTXO set as well as a set of dangling transactions. However, in the coordinated version of the protocol, there can't be any dangling transactions and thus, it is no longer required to check applicability of those transactions to the UTXO set. It suffices to store a hash of the resulting outputs of that UTXO instead.

checkClose :: ScriptContext -> HeadContext -> [Party] -> SnapshotNumber -> [Signature] -> Bool

txOutAdaValue :: TxOut -> Integer

txInfoAdaFee :: TxInfo -> Integer

checkFanout :: BuiltinByteString -> Integer -> ScriptContext -> Bool

(&) :: a -> (a -> b) -> b

mustBeSignedByParticipant :: ScriptContext -> HeadContext -> Bool

matchParticipationToken :: CurrencySymbol -> Value -> [TokenName]

mustContinueHeadWith :: ScriptContext -> Address -> Integer -> Datum -> Bool

txOutDatum :: TxInfo -> TxOut -> Datum

hashPreSerializedCommits :: [SerializedTxOut] -> BuiltinByteString

hashTxOuts :: [TxOut] -> BuiltinByteString

verifySnapshotSignature :: [Party] -> SnapshotNumber -> [Signature] -> Bool

verifyPartySignature :: SnapshotNumber -> Party -> Signature -> Bool

mockVerifySignature :: Party -> SnapshotNumber -> BuiltinByteString -> Bool

mockSign :: Integer -> BuiltinByteString -> BuiltinByteString

typedValidator :: TypedValidator Head

The script instance of the auction state machine. It contains the state machine compiled to a Plutus core validator script. TODO: Add a NetworkId so that we can properly serialise address hashes see encodeAddress for details

compiledValidator :: CompiledCode (ValidatorType Head)

validatorHash :: ValidatorHash

address :: Address

validatorScript :: Script

Get the actual plutus script. Mainly used to serialize and use in transactions.