Hydra.Contract.Head

Documentation

data State

Constructors

Initial ContestationPeriod [Party]

Open

Final

data Input

Constructors

CollectCom

Abort

type Head = StateMachine State Input

hydraStateMachine :: MintingPolicyHash -> StateMachine State Input

hydraTransition :: State State -> Input -> Maybe (TxConstraints Void Void, State State)

typedValidator :: MintingPolicyHash -> TypedValidator (StateMachine State Input)

The script instance of the auction state machine. It contains the state machine compiled to a Plutus core validator script. The MintingPolicyHash serves two roles here:

Parameterizing the script, such that we get a unique address and allow for multiple instances of it

Identify the 'state thread token', which should be passed in transactions transitioning the state machine and provide "contract continuity"

validatorHash :: MintingPolicyHash -> ValidatorHash

address :: MintingPolicyHash -> Address

machineClient

Arguments

:: AssetClass
PolicyId for the head instance

-> StateMachineClient State Input

The machine client of the hydra state machine. It contains both, the script instance with the on-chain code, and the Haskell definition of the state machine for off-chain use.