Hydra.Contract.Head

Documentation

type DatumType = State

type RedeemerType = Input

hydraHeadV1 :: BuiltinByteString

headValidator :: State -> Input -> ScriptContext -> Bool

data CheckCollectComError

Constructors

NoContinuingOutput

MoreThanOneContinuingOutput

OutputValueNotPreserved

OutputHashNotMatching

data HeadContext

Constructors

HeadContext

Fields

headAddress :: Address

headInputValue :: Value

headCurrencySymbol :: CurrencySymbol

mkHeadContext :: ScriptContext -> HeadContext

checkAbort :: ScriptContext -> HeadContext -> [Party] -> Bool

On-Chain verification for Abort transition. It verifies that:

All PTs have been burnt: The right number of Head tokens, both PT for parties and thread token, with the correct head id, are burnt,

All committed funds have been redistributed. This is done via v_commit and it only needs to ensure that we have spent all comitted outputs, which follows from burning all the PTs.

checkCollectCom

Arguments

:: ScriptContext
Script execution context

-> HeadContext
Static information about the head (i.e. address, value, currency...)

-> (ContestationPeriod, [Party])
Initial state

-> Bool

On-Chain verification for CollectCom transition. It verifies that:

All participants have committed (even empty commits)

All commits are properly collected and locked into the contract as a hash of serialized tx outputs in the same sequence as commit inputs!

The transaction is performed (i.e. signed) by one of the head participants

It must also initialize the on-chain state η* with a snapshot number and a hash of committed outputs.

(*) In principle, η contains not a hash but a full UTXO set as well as a set of dangling transactions. However, in the coordinated version of the protocol, there can't be any dangling transactions and thus, it is no longer required to check applicability of those transactions to the UTXO set. It suffices to store a hash of the resulting outputs of that UTXO instead.

checkClose :: ScriptContext -> HeadContext -> [Party] -> BuiltinByteString -> SnapshotNumber -> BuiltinByteString -> [Signature] -> ContestationPeriod -> Bool

The close validator must verify that:

The closing snapshot number and signature is correctly signed

The resulting closed state is consistent with the open state or the closing snapshot, depending on snapshot number

The transaction is performed (i.e. signed) by one of the head participants

makeContestationDeadline :: ContestationPeriod -> ScriptContext -> POSIXTime

Checks that the datum

checkContest

Arguments

:: ScriptContext

-> HeadContext

-> POSIXTime

-> [Party]

-> SnapshotNumber
Snapshot number of the closed state. XXX: Having two snapshot numbers here is FRAGILE

-> SnapshotNumber
Snapshot number of the contestin snapshot.

-> BuiltinByteString

-> [Signature]

-> Bool

The contest validator must verify that:

The contest snapshot number is strictly greater than the closed snapshot number.

The contest snapshot is correctly signed.

The resulting closed state is consistent with the contested snapshot.

The transaction is performed (i.e. signed) by one of the head participants

checkHeadOutputDatum :: ToData a => ScriptContext -> a -> Bool

txOutAdaValue :: TxOut -> Integer

txInfoAdaFee :: TxInfo -> Integer

checkFanout :: BuiltinByteString -> POSIXTime -> Integer -> ScriptContext -> Bool

(&) :: a -> (a -> b) -> b

mustBeSignedByParticipant :: ScriptContext -> HeadContext -> Bool

findParticipationTokens :: CurrencySymbol -> Value -> [TokenName]

mustContinueHeadWith :: ScriptContext -> Address -> Integer -> Datum -> Bool

findTxOutDatum :: TxInfo -> TxOut -> Datum

hashPreSerializedCommits :: [Commit] -> BuiltinByteString

Hash a potentially unordered list of commits by sorting them, concatenating their preSerializedOutput bytes and creating a SHA2_256 digest over that.

hashTxOuts :: [TxOut] -> BuiltinByteString

Hash a pre-ordered list of transaction outputs by serializing each individual TxOut, concatenating all bytes together and creating a SHA2_256 digest over that.

verifySnapshotSignature :: [Party] -> SnapshotNumber -> BuiltinByteString -> [Signature] -> Bool

verifyPartySignature :: SnapshotNumber -> BuiltinByteString -> Party -> Signature -> Bool

compareRef :: TxOutRef -> TxOutRef -> Ordering

compiledValidator :: CompiledCode ValidatorType

validatorScript :: Script

validatorHash :: ValidatorHash