Hydra.Contract.Initial

Description

The initial validator which allows participants to commit or abort.

Documentation

data InitialRedeemer

Constructors

ViaAbort

ViaCommit

Fields

committedRefs :: [TxOutRef]

Points to the committed Utxo.

type DatumType = CurrencySymbol

type RedeemerType = InitialRedeemer

validator

Arguments

:: ScriptHash
Hash of the commit validator

-> DatumType

-> RedeemerType

-> ScriptContext

-> Bool

The v_initial validator verifies that:

spent in a transaction also consuming a v_head output

ensures the committed value is recorded correctly in the output datum

ensures that the transaction was signed by the key corresponding to the PubKeyHash encoded in the participation token name

NOTE: It does not need to ensure that the participation token is of some specific Head currency.

checkCommit

Arguments

:: ScriptHash
Hash of the commit validator

-> CurrencySymbol
Head id

-> [TxOutRef]

-> ScriptContext

-> Bool

compiledValidator :: CompiledCode ValidatorType

validatorScript :: SerialisedScript

validatorHash :: ScriptHash

datum :: DatumType -> Datum

redeemer :: RedeemerType -> Redeemer