Hydra.Contract.Initial

Description

The initial validator which allows participants to commit or abort.

Documentation

data InitialRedeemer

Constructors

ViaAbort

ViaCommit

Fields

committedRef :: Maybe TxOutRef

Points to the committed Utxo.

type DatumType = ()

type RedeemerType = InitialRedeemer

validator

Arguments

:: ValidatorHash
Commit validator

-> ()

-> InitialRedeemer

-> ScriptContext

-> Bool

The v_initial validator verifies that:

FIXME: spent in a transaction also consuming a v_head output

ensures the committed value is recorded correctly in the output datum

ensures that the transaction was signed by the key corresponding to the PubKeyHash encoded in the participation token name

NOTE: It does not need to ensure that the participation token is of some specific Head currency.

checkAuthor :: ScriptContext -> Bool

Verifies that the commit is only done by the author

checkCommit

Arguments

:: ValidatorHash
Commit validator

-> Maybe TxOutRef

-> ScriptContext

-> Bool

debugInteger :: Integer -> BuiltinString

Show an Integer as decimal number. This is very inefficient and only should be used for debugging.

compiledValidator :: CompiledCode ValidatorType

validatorScript :: Script

validatorHash :: ValidatorHash

datum :: DatumType -> Datum

redeemer :: RedeemerType -> Redeemer