Hydra.Contract.Util

Documentation

hydraHeadV1 :: BuiltinByteString

hasST :: CurrencySymbol -> Value -> Bool

Checks that the output contains the state token (ST) with the head CurrencySymbol and TokenName of hydraHeadV1

mustBurnAllHeadTokens :: Value -> CurrencySymbol -> [Party] -> Bool

Checks all tokens related to some specific CurrencySymbol.

This checks both PTs and ST are burnt.

mustBurnST :: Value -> CurrencySymbol -> Bool

Checks if the state token (ST) for list of parties containing specific CurrencySymbol are burnt.

mustNotMintOrBurn :: TxInfo -> Bool

(===) :: Value -> Value -> Bool infix 4

Checks for exact equality between two serialized values. Equality on value is very memory intensive as it's defined on associative lists and AssocMap equality is implemented. Instead we can be more strict and require EXACTLY the same value and compare using the serialised bytes.

Errors

data UtilError

Constructors

MintingOrBurningIsForbidden